// frontend/src/app/models/geological-interests.model.ts
export interface GeologicalInterests {
    stratigraphic: boolean;
    paleontological: boolean;
    sedimentological: boolean;
    geomorphological: boolean;
    mineral_resource: boolean;
    mineralogical: boolean;
    metamorphism: boolean;
    volcanism: boolean;
    plutonism: boolean;
    hydrogeology: boolean;
    tectonics: boolean;
  }